+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Changelog
 * Runtime: added the ability to override render settings in :doc:`boot.config <reference/boot_config>`.
 * Runtime: added the ability to specify a configuration file with user-specific settings via the ``user_config`` setting in :doc:`boot.config <reference/boot_config>`.
 * Runtime: Linux: it is now possible to run with either OpenGL or Vulkan rendering backends.
+* Runtime: added ``physics`` console command to control debug visualizations.
 * Lua: added ``GameSave`` singleton to save/load savegames.
 
 **Fixes**